Feature/sonar compatibility

+++ test/oap2/format/OAR016/fail-example.js
@@ -15,7 +15,7 @@ module.exports = {
               "properties": {
                 "amount": {
                   "type": "number",
-                  "format": "decimal"
+                  "format": "int32"
                 },
                 "nested": {
                   "type": "object",
